implementing test cases for UserService

+++ src/main/java/com/ios/service/UserService.java
@@ -25,18 +25,21 @@ public class UserService {
 		return userRepository.save(user);
 	}
 
-	public ResponseEntity<String> deleteUserById(int userId) {
+	public ResponseEntity<String> deleteUserById(int userId) throws UserNotFoundException {
+		if (!userRepository.existsById(userId)) {
+			throw new UserNotFoundException("User Not Found With USerId : " + userId);
+		}
 		userRepository.deleteById(userId);
 		return new ResponseEntity<>("User Deleted..!", HttpStatus.NO_CONTENT);
 	}
 
 	public User updateUser(User user, int userId) throws UserNotFoundException {
-		User foundUser = userRepository.findById(userId)
-				.orElseThrow(() -> new UserNotFoundException("User Not Found With USerId : " + userId));
+		if(!userRepository.existsById(userId)) {
+			throw new UserNotFoundException("User Not Found With USerId : " + userId);
+		}				
 
-		foundUser.setUerName(user.getUerName());
-		foundUser.setEmail(user.getEmail());
-		return userRepository.save(foundUser);
+		user.setUserId(userId);
+		return userRepository.save(user);
 
 	}
